RtpsRelay double-sends to application participant
Problem
-------

Suppose a participant has been discovered and one
publication/subscription has been discovered.  Suppose that the
participant then sends a directed message to the application
participant.  The vertical handler will flag this message as necessary
to forward to the application participant.  However, since the sending
participant has a partition, the "to" list will also be evaluated and
result in a send to the application participant.

Solution
--------

Remove the application participant from the "to" list.