Enums do not have extensibility set
Problem
-------

The IDL compiler does not set the extensibility for an enum.

Solution
--------

Set the extensibility in the IDL compiler to appendable if there is no
annotation.  For backwards compatibility, 1) the IDL compiler will have an
option for not setting the extensibility and 2) a zero extensibility
will be treated as a wild-card in the assignability check.