Repo contains references to Redmine
Problem
-------

The OpenDDS repository contains references to Redmine which is an
issue tracker that is no long used.  Specifically, the code references
issue 1446.  A log message for this issue is treated as an error.
The current consensus is that this is not an error.

Solution
--------

Remove the references to Redmin in the log message and tests that
ignore the corresponding log message.