Rewrite DataWriterImpl[_T] to use AbstractSample
This moves a lot of the functionality of DataWriterImpl_T into
DataWriterImpl by using a new class called AbstractSample to avoid
working with the native C++ IDL-generated types directly as much as
possible. Some more work is needed on this front though. The ultimate
goal of this class is to be able to use it with DynamicData as well as
with the native types to work towards implementing a DynamicDataWriter
and DyanmicDataReader.