Fix Configure Submodule Checkout on Windows
If doing a default build on Windows with `configure` in the Visual
Studio Developer Command Prompt and no git installed on the PATH, the
configure script's `which` function will erroneously report there is a
git command available and try, but fatally fail, to use git to get the
commit hash needed for RapidJSON.

which is erroneous because it is using `-x`, which seems to return true
for directories on Windows. There is a `Git` directory in the PATH that
Visual Studio adds that it's picking up. As a side note this actually
has a working git command in it, but since it's not on the PATH I'm not
sure if we should try to take advantage of it.