Add a `mwc.pl` Wrapper Script
This script uses the same `mwc.pl` arguments (minus the workspace file)
that OpenDDS was configured with. This is helpful for cases where
ACE/TAO was separately configured in a generic way. In this case the
default features file doesn't get updated and each feature has to be
passed if running `mwc.pl` manually. This script will contain the
features as well as the `-type` and `-static` if needed.