Quick shutdown leads to error in LinuxNetworkConfigMonitor
Problem
-------

As part of startup, the Service Participant opens the socket
for the LinuxNetworkConfigMonitor and then uses the ReactorInterceptor
to register it with the reactor.  As part of shutdown, the Service
Participant closes the socket for the LinuxNetworkConfigMonitor and
then uses the ReactorInterceptor to unregister.  If the shutdown
happens immediately after startup, then the reactor thread will see a
closed socket, i.e., the order of operations is:

1. open
2. close
3. register -> error, invalid handle

Solution
--------

Defer all of the logic to the ReactorInterceptor commands.

Note
----

The NetworkConfigMonitor interface expects a synchronous open and
close.  It does this so it can fall back from the
LinuxNetworkConfigMonitor to other implementations.  To satisfy this
while ensuring the correct order of operations, conditions were added
to the ReactorInterceptor commands.  In general, this is dangerous but
none of the alternatives for fixing it are attractive at this point:

* Force it upon the user, i.e., the user is responsible for writing to
  the `network_interface_address_topic`
* Replacing ACE