Move (Ref)Wrapper, Add special_serialization Anno.
RefWrapper is going to be needed for dynamic_data_adapter_generator, so
I moved it from marshal_generator to dds_generator. The only problem
with this is it needs to know if a type has special serialization. To
make that work independently of marshal_generator I changed how it works
from name-based to one based on a new annotation:
@OpenDDS::internal:::special_serialization.