Add Release Guide to Sphinx Documentation
This has all OCI-internal parts removed (I hope) and SFTP and Doxygen
parts removed as we don't know if and how that's going to be involved
for now. Announcement section was also removed.

There are other more minor changes throughout the document, particularly
merging and tweaking the parts that are now in the "Update Files in the
Repo as Needed".