TcpDataLink fails to send association message
Problem
-------

The `send_association_msg` method creates a special `REQUEST_ACK`
control message to signal that the local reader is ready to receive
data.  On receipt of the message, the remote writer completes the
association and can start sending data.  `send_association_msg` calls
`send_i` which checks if the publication id of the
`TransportControlElement` is for a stopped (local) client.  The
publication id returned by the `TransportControlElement` is always
GUID_UNKNOWN.  Thus, if GUID_UNKNOWN gets put into the stopped client
set then the `TcpDataLink` will never send the association message and
association will stall on the writer side.

`TransportClient` caches its guid when an association is made.
Suppose the `TransportClient` is stopped before an association.  In
this case the guid is `GUID_UNNOWN` and this guid is put in the
stopped client set of the `TcpDataLink`.

Solution
--------

1. Implement support for the publisher id.
2. Guard against usages of an uninitialized `repo_id_` in `TransportClient`.