Add Checks for RTPS Unicast Ports
The loop for opening RTPS ports for a participant could theoretically go
on forever. This stops it once it goes past the max UDP port. Also add a
warning for going past the max participant id allowed in the RTPS spec
(119, in section 9.6.2.3). As long as it doesn't result in an invalid
port, going past that shouldn't cause a problem, but it could be a sign
that OpenDDS has a limited number of ports at its disposal. Also another
implementation could use this as a hard limit, but that's much less of a
concern.