OpenDDS lacks a dynamic configuration store
Problem
-------

Configuration in OpenDDS has the following issues:

* Values that can be set via configuration file cannot be set via
  command line.
* Values that can be set via API cannot be set via configuration file.
* Configuration APIs are written in C++ and not fully exposed for
  other language bindings.
* Dynamic reconfiguration is messy.
* OpenDDS cannot be configured via environment variables.
* OpenDDS cannot take advantage of multiple configuration files.

Solution
--------

Implement a store of key-value pairs based on an Internal DDS Topic.
The key is the canonicalized name of the configuration value and the
value is the desired value stored as a string.  Using an Internal DDS
Topic will allow different modules to subscribe to the configuration
for dynamic updates.

The interface to the configuration store will be captured in IDL so
that it can be exposed for different language bindings.

Subsequent work will apply the store to command-line arguments,
configuration file processing, and environment variables.