Tweaks News Docs, Remove Auto Inserting PR Numbers
It would be convenient, but it might cause problems. For example when
gathering commits for a patch release, if only the actual PR commits are
considered, then the auto commits could be left out. For now PR numbers
will either have to be predicted or pushed after creating the PR.