TransportClient uses uninitialized GUID
Problem
-------

The TransportClient class does not initialize its guid until an
association.  However, the client can and is used before an
association.  One example is a DataWriter that sends liveliness
messages.  The fix introduced by #4120 caused various functions to
short-circuit when the guid was not set.  However, this introduced a
deadlock in the example where the DataWriter was waiting for the
liveliness message to be delivered.

Solution
--------

Three alternatives were considered.

First, extending the solution in #4120 lacked confidence as it was
impossible to explore all of the corner cases.

Second, the guid was passed in the constructor of the TransportClient.
This solution failed as the current design in concert with
spec-defined behavior 1) allows DataWriters and DataReaders to be
created before enabling them and 2) determines the guid during the
call to enable via a call into discovery.  While this solution was the
best in terms of maintainability, it would require extensive
refactoring.

Third, modifying discovery to set the guid as soon as it is known and
before any associations are processed.  This solution was close to the
second one in terms of the class invariant of an initialized guid
while fitting within the current design.

Discovery implementations now how the responsibility of calling
`set_publication_id` and `set_subscription_id` when a publication or
subscription is added.  DataReaderImpl was modifying to not use the
TransportClient when the transport is disableb as it is for built-in
topics.