RTPS Parameter parsing erroneously succeeds when size is 0
Problem
-------

The specialized RTPS Parameter parsing code in `marshal_generator.cpp`
accepts parameters that declare a zero size by setting the
discriminator in the union and returning successfully.  If the
discriminator is for a declared case of the union, then this creates a
problem since the corresponding member of the union is not
initialized.  This is especially problematic for members that require
dynamic allocation like strings and sequences since the code accessing
the Parameter will see a null pointer.

Solution
--------

Remove the special handling for size==0.  This causes known cases to
be parsed.  If size==0 and the discriminator is a known case, then the
parse will fail.