Shapes demo is lumped with quickstarts
Problem
-------

The Shapes demo allows users to download an OpenDDS application and
use it.  In that case, it is a quick start.  However, the intent of
the demo is to explore different QoS settings and DDS features.

Solution
--------

Move the Shapes demo to after the introduction to DDS as a
continuation of the exploration of DDS concepts.