"Un-cache" Feature Lists in CMake
Fixes #4328

The equivalent to many MPC features in CMake are built up using a
function called `_opendds_feature` that adds features to lists to use.
The issue is these lists are set as cache variables, which are
persistent across CMake runs. This resulted in the features being
appended each time, so they are listed multiple times in output and
`default.features` if building ACE.

These don't really have to be cache variables, so just set them as
normal variables in the parent scope.

Also fixed some typos in the same file and added a note about removing
an include in the future.