Publish/subscribe actions do not have a validity section
Problem
-------

In a collaborative DDS Domain, such as those possible with the DDS
Permissions Manager, different organization will contribute different
publish/subscribe actions for the same application (subject).  These
publish/subscribe actions will be valid in different timeframes.

Solution
--------

A desirable solution would be to have a permissions document with
multiple grants referring to the same application.  However, the
"default action" in a DDS Security grant effectively means that an
application (subject) can have a single grant.  Thus, all of the
publish/subscribe actions for an application must be combined under a
single grant.  This means that either 1) the validity for the grant
must be crafted carefully from the validity assigned to each action or
2) each action needs a validity interval.

Crafting the overall grant validity is possible but can lead to
pathological behavior with respect to renewing a permissions document.
That is, the application must new the permissions document at every
point in time (not_before or not_after) for every publish/subscribe
action.  To avoid this complexity, each publish/subscribe action is
given a validity which is independent from the grant validity.

This is an OpenDDS extension to DDS Security.