index is not generated for GHA builds
Problem
-------

The index.html is not generated for the GHA builds.  This is
preventing the scoreboard from showing these builds.

Solution
--------

Generate the index.

Notes
-----

Generating the index this way will only allow the most recent build to
appear on the scoreboard.  Adding previous builds to the index will be
addressed later.