Transport templates are not reused
Problem
-------

One of the goals of using the ConfigStore (#4134) is to minimize the
number of entries in the ConfigStore that are generated by the
framework itself.  Such entries are basically leaked.

This commit removes the need to instantiate `[rtps_discovery]` sections.

Solution
--------

Parameterize `RtpsUdpTransportInst::multicast_group` on the
domain and participant so that it can be customized on the fly.