Fix Broken PDF Scraping in Sphinx
This might be caused by a refactor that PyMuPDF is currently doing. This
library is used to read the OMG spec table of contents to get the links.
I could look into this further to see if this was intentional or a bug,
but the fix seems to be trivial.