Fix for generating std::nullopt when it shouldn't
type_to_default, in more places we generate with is_optional than we do
is_union, so I placed is_optional first in the function definition this
broke it in a single place. Will need to decide which is easier to get,
the is_optional from the generateCaseBody or the is_union from the other
places we call type_to_default.