OpenDDS cannot be configured with environment variables
Problem
-------

OpenDDS cannot be configured with environment variables.  Configuring
with environment variables is useful for containers and cloud
deployments.

Solution
--------

Parse environment variables and add them to the ConfigStore.  Only
environment variables starting with OPENDDS_ are parsed.  The
remainder of the name is used as the key for the ConfigStore.
Environment variables are processed before command-line arguments and
the config file.