Internal DDS cannot efficiently filter on instances
Problem
-------

It is foreseeable that "demuxing" will be a necessary feature of the
Internal DDS system.  For example, suppose a socket abstract writes
RTPS message samples which are keyed by the participant guid prefix.
A reader is probably only interested in one participant.  The write
could give the sample to all readers which would be inefficient since
only one reader is actually interested.

Solution
--------

This solution assumes that an InternalDataReader reader is interested
in a set of instances that known a priori and are specified
explicitly.  The InternalDataReader indexes attached readers on the
instances to efficiently deliver samples to readers that are only
interested in those instances.

Note
----

A "real" DDS systems would content-filtering to accomplish the same
objective.