Moved RTPS SubmessageKind to its own IDL file to use the value annotation
Includes limited support for IDL4 "standardized" value annotation.
When opendds_idl is used to generate the language mapping for an enum,
an enumerator annoatated with value() will be assigned that value.

Also changes opendds_idl options used for IDL internal to the RTPS library
in order to minimize footprint of generated code.