`--mpcopts` values are split on spaces
Problem
-------

`--mpcopts` is a `configure` script option that accumulates additional
options for MPC.  These values are split on spaces so
`--mpcopts='-value_template build_flags+=-Wall -Werror'` results in three
arguments to MPC: `-value_template`, `build_flags+=-Wall` and
`-Werror`.  This causes the configure script to fail because the
arguments to MPC are incorrect.

Solution
--------

Remove the splitting functionality.

The examples above would now be written as
* `--mpcopts=-value_template --mpcopts=build_flags+="-Wall -Werror"`

Introduce a new option `--mpc:opt arg`.  The examples would then be
* `--mpc:value_template build_flags+="-Wall -Werror"`