Inconsistent setting of `OPENDDS_SECURITY` corrupts builds
Problem
-------

`OPENDDS_SECURITY` is a macro that appears in public headers.
If a user enables security during the build of OpenDDS but then does
not set it during the build of their application, then the build is
inconsistent and the application fails in unexpected ways.  Regardless
of the build system (MPC, CMake, or something else), the idea that
security is enabled must be preserved to generate consistent builds.

Solution
--------

See https://github.com/OpenDDS/OpenDDS/issues/4616 for background.

Add `OPENDDS_CONFIG_SECURITY` to `dds/OpenDDSConfig.h`.

Add `dds/OpenDDSConfigWrapper.h` and `dds/OpenDDSConfigWrapper.idl`
which accommodate both `OPENDDS_SECURITY` and
`OPENDDS_CONFIG_SECURITY`.