Update char_literals to Check for wchar Esc. Seqs
https://github.com/DOCGroup/ACE_TAO/pull/2244 sets
`TAO_IDL_HAS_OCTET_AND_WCHAR_UNION_DISCS` to 1, but the char_literals
wchar union contains all the escape sequences literals, which are not
supported in TAO_IDL: https://github.com/DOCGroup/ACE_TAO/issues/1284.
This changes it to check the planned macro name.

Also removed the hex and octet escape sequences for wchar, which are
probably not indented to work with wchar.