Removed the "expand" section of rapidjson.mpb
Expanding the variable here is useful because it provides a default
value when RAPIDJSON_ROOT is not set in the environment.

However, it always results in makefiles that hard-code the expansion
value (absolute path to rapidjson, may be in DDS_ROOT).

With this change, if the MPC feature no_rapidjson is 0 then the
RAPIDJSON_ROOT environment variable needs to be set either by
the configure script or manually.