Fixed Broken Links, TypeLookup Sec Entity Names
There were links that should have gone to
https://issues.omg.org/issues/DDSSEC12-86 that that were wrong. On
inspection, the issue was fixed in v1.2 of the DDS security spec, which
just came out. It looks like what we have matches what's in 1.2, so I
removed the links.

I also switched the "reader"/"writer" and "secure" in names because
it's what's in 1.2 and matches the other builtin entities.