RtpsRelay tests don't have CheckSourceIp=false
Problem
-------

The RtpsRelay tests are a starting poing to users that want to use the
RtpsRelay.  In any realistic deployment, the user must have
`CheckSourceIp=false` in the configuration file.  Since the tests
don't have this line, the user will have to learn "the hard way."

Solution
--------

Add `CheckSourceIp=false` to the configuration files.