OpenDDS version not accounted for in scoreboard
Problem
-------

As we approach OpenDDS 4, we will most likely have CI jobs with the
same name running against version 3 and 4.  Currently, the OpenDDS
version is not used when naming the result.  For clarity moving
forward, the result should include the version.

Solution
--------

Extract the OpenDDS version from `dds/Version.h` and use it in the
upload path.