Fixed Leak in AddressCache for RTPS/UDP Transport
AddressCache has a `map_` that maps keys (local-remote GUID pairs in
RTPS/UDP) to cache entries and a `id_map_` that maps GUIDs to a vector
of related keys. When a key is inserted into the cache the GUIDs get get
entries in the `id_map_` that can be used to search `map_`.

`id_map_` has two problems:
- The vector of keys are not checked if they are being duplicated. These
  changes make it a set.
- `remove_id` removes the `id_map_` keys for the GUID argument (the
  remote GUID in RTPS/UDP), but not any keys that contain that GUID for
  other vectors in `id_maps_` (the remote in RTPS/UDP). These changes
  try to clean up all the revelant GUIDs.

Also:
- A change to release reservation I meant to add in the other leak PR.
  If the release reservation is for a specfic local-remote pair, it
  doesn't seem right to remove all messages for the remote from the send
  queue.
- Small fixes to QoS doc that I had in this local repo for some reason.