Fix OMG Spec Links Page
[This page](https://opendds.readthedocs.io/en/master/internal/omg_spec_links.html)
was causing a lot of output in the link checker, so I made it
conditional in https://github.com/OpenDDS/OpenDDS/pull/4669.
Unfortunately it looks like I also broke the page so it never lists all
the possible spec section links.

Also added `--nitpicky', `--keep-going', and `--show-traceback' to the
`sphinx-build` command of `build.py strict` to make output more useful.