Updated ThreadStatusManager for iterative event handling
Some event loop process entire batches of events at once.
Before this change, the batch of data would be considered an "Event"
for the ThreadStatusManager.  With this change, each element within
the batch counts as its own event for the ThreadStatusManager.
That way the thread appears to be active and not stuck.