Thread monitors may be starved for data
Problem
-------

When thread status monitoring is enabled, all participants that use
RTPS discovery will harvest the thread status from the monitor and
publish it on the OpenDDSInternalThread BIT.  If there are multiple
participants, they will each harvest.  Typically, an application will
only look at one of the BIT readers with the result being that any
monitoring logic may be operating on an incomplete set of data.

Solution
--------

Add a property `RTPS_HARVEST_THREAD_STATUS` that enables harvesting
for particular participants.