Service Participant thread doesn't update thread status
Problem
-------

The Service Participant thread invokes the BIT listeners in the
RtpsRelay.  Given large enough batches, the listeners can occupy the
processor long enough for the thread monitoring logic to trip.

Solution
--------

Add events for each sample to update the thread status.

Add logging to show the size of the batches.