Isolate `opendds_target_sources` Code Generation
This was prompted by build failures in Ninja with CMake 4.0 in
https://github.com/OpenDDS/OpenDDS/pull/4959 caused by setting the CMake
"max policy version" from 3.27 to 4.0. They seem to be related to
https://cmake.org/cmake/help/latest/policy/CMP0154.html. CMake is trying
to optimize the Ninja build by assuming generated files are "private" by
default in targets using file sets so it can defer generating them.
OpenDDS built using CMake does use file sets, but it doesn't use file
sets for IDL-generated files, so Ninja might attempt to build source
files that use IDL-generated headers before those headers are generated.

We could use `target_sources(... FILE_SET ...)` within
`opendds_target_sources` to create a file set for generated headers, but
it would be problematic:
- It requires the file set to be installed using
  `install(TARGETS ... FILE_SET ...)` and this is something the user has
  to do because seems like it can only be called once per target.
- Not only can we not do it for them, but they would be forced into
  doing it by CMake, even if they don't want to setup installation for
  the target at that moment or at all.
- Right now the design of `opendds_target_sources` is explicitly trying
  to be flexible by being agnostic about the installation method of
  generated files and our documentation explicitly says this.

The workaround for this is to create separate custom targets for each
IDL file that have to be executed first.
As part of this I went ahead with
https://github.com/OpenDDS/OpenDDS/issues/4306 (`opendds_target_sources`
Should Generate Files Independently Per-Target Instead of
Per-CMakeLists) creating separate custom targets probably makes sharing
the IDL command target even harder than it was before.

This all allows the "max policy version" to be set for 4.0.

Also:
- Move `include(GNUInstallDirs)` to `opendds_util.cmake` for
  `opendds_install_interface_files`.
- Support CMake's `codegen` target to compile IDL when CMake supports
  it: https://cmake.org/cmake/help/latest/policy/CMP0171.html
- `opendds_target_sources` shouldn't automatically generate an export
  header there's only `PRIVATE` IDL files.