RtpsRelay leaks memory for incomplete discoveries
Problem
-------

Observations:

1. `GuidPartitionTable::lookup` creates a cache entry if it does not
   exist.
2. The discovery of a publication or subscription removes
   the cache entry if it exists and the partitions change.
3. The removal of a publication or subscription removes the cache
   entry if it exists.
4. `GuidPartitionTable::lookup` can be called before discovery
   completes.

In a scenario where discovery completes, the cache is eventually
correct, that is, it is correct after the first lookup that follows
the last publication or subscription.

In a scenario where dicovery doesn't complete, then the cache entry is
created (1 and 4) but not removed (2 and 3).

Solution
--------

Only create a cache entry if the set of partitions is not empty which
implies discovery is in progress or has completed.