GHA Annotations and Better Link Check for Sphinx
This should make Sphinx output the correct annotation in RST files in
the PR in most situations except for when the content causing the error
is in a news fragment. News fragments are used to generate two files at
runtime, so the source info that Sphinx has for these files doesn't
create the correct annotation in the PR.

Also tried to improve link checking so it fails less often. This
includes:

- Ignoring `:ghfile:` links. It's not really needed because they're
  checked locally. There's also a lot of them so it takes a while to
  check them and the there's a greater chance for something going wrong.
- Ignoring links that fail with 409 HTTP status code. They happen a lot.
- Turning redirects on GitHub PR and issue links into errors because
  that means something is wrong with the link.