Heartbeat fallback is not reset for new samples
Problem
-------

Suppose there is a reliable writer with two readers A and B at steady
state.  Reader B goes away without disposing/unregistering itself. The
writer then writes a sample, sends a heartbeat, and reader A
acknowledges. Since reader B does not acknowledge, the writer will
continue to send heartbeats with an increasing interval between
them (fallback). When A writes again, the fallback is not reset. In
the event that the sample was lost, it will take the current fallback
between the writer sends a heartbeat and reliability repairs the
missing sample.

Solution
--------

Reset the heartbeat fallback when samples are written or the sequence
number advances in other ways.